refactor(runtime): split workspace file resolution into focused modules
